Move everything to the Pusher namespace
