+++ lib/pusher-signature.rb
@@ -1,0 +1,1 @@
+require 'pusher/signature'
