+++ lib/pusher/signature/version.rb
@@ -1,0 +1,5 @@
+module Pusher
+  module Signature
+    VERSION = "0.1.8"
+  end
+end
