--- lib/signature/version.rb
@@ -1,3 +1,0 @@
-module Signature
-  VERSION = "0.1.8"
-end
