+++ spec/spec_helper.rb
@@ -1,9 +1,9 @@
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
-require 'signature'
+require 'pusher-signature'
 
 require 'rspec'
 require 'em-spec/rspec'
 
 RSpec.configure do |config|
-  
+
 end
